+++ src/lua.rs
@@ -1,5 +1,8 @@
+use crate::config::Config;
 use anyhow::bail;
 use anyhow::Result;
+use mlua::Lua;
+use mlua::LuaSerdeExt;
 
 pub fn resolve_fn<'lua, 'a>(
     table: &mlua::Table<'lua>,
@@ -14,4 +17,48 @@ pub fn resolve_fn<'lua, 'a>(
     } else {
         bail!("Invalid path")
     }
+}
+
+// TODO: Remove config input when config.yml is deprecated
+pub fn init(lua: &Lua, lua_script: &str, config: Config) -> Result<Config> {
+    let globals = lua.globals();
+
+    let lua_xplr = lua.create_table()?;
+    lua_xplr.set("config", lua.to_value(&config)?)?;
+
+    let lua_xplr_fn = lua.create_table()?;
+    let lua_xplr_fn_builtin = lua.create_table()?;
+    let lua_xplr_fn_custom = lua.create_table()?;
+
+    lua_xplr_fn.set("builtin", lua_xplr_fn_builtin)?;
+    lua_xplr_fn.set("custom", lua_xplr_fn_custom)?;
+    lua_xplr.set("fn", lua_xplr_fn)?;
+    globals.set("xplr", lua_xplr)?;
+
+    lua.load(&lua_script).set_name("init")?.exec()?;
+
+    let version: String = match globals.get("version").and_then(|v| lua.from_value(v)) {
+        Ok(v) => v,
+        Err(_) => bail!("'version' must be defined globally in init.lua"),
+    };
+
+    let lua_xplr: mlua::Table = globals.get("xplr")?;
+
+    let config: Config = lua.from_value(lua_xplr.get("config")?)?;
+    Ok(config.with_version(version))
+}
+
+pub fn extend(lua: &Lua, lua_script: &str) -> Result<Config> {
+    lua.load(&lua_script).set_name("init")?.exec()?;
+
+    let globals = lua.globals();
+    let version: String = match globals.get("version").and_then(|v| lua.from_value(v)) {
+        Ok(v) => v,
+        Err(_) => bail!("'version' must be defined globally in init.lua"),
+    };
+
+    let lua_xplr: mlua::Table = globals.get("xplr")?;
+
+    let config: Config = lua.from_value(lua_xplr.get("config")?)?;
+    Ok(config.with_version(version))
 }
